Add utility to print the forwarding table entries.
The current table display utility uses kernel printk instead of sysfs
file io. The problem is printk output does not belong to the output pipe
of the shell that initializes the query. The other thing is that the
output format is subject to Marvell's code change. This CL provides the
utility to print forwarding tables in python friendly format.

Three new files are added:
/sys/devices/platform/tpm/cfg_l2/show_l2_rule
/sys/devices/platform/tpm/cfg_l2/show_l2_key
/sys/devices/platform/tpm/cfg_frwd/show_frwd_rule

By opening these three files, you will see the latest forwarding
information related to media converter configuation.

This CL does not export every table since many of those are not related
to media converter. But through this same framework, we can expand them
easily when they are needed.

Change-Id: I3ce8d47729562fe191b5b93aca6785474acd0674