Hold the lock for longer when copying the kernel dmesg buffer.
This function would acquire and release a spinlock for every single byte.
If you used a large buffer size like 8MB, this could make the syslog()
function take more than a second to copy the data.  On our device it was 1.5
seconds; with this change it's reduced to 0.17 seconds.

Change-Id: If63befbe18e4fc215f2e89f5c02dce87ea9064f2

Conflicts:
	kernel/printk.c