Add support to detect link change.
The current Marvell link state update is dependent on the state machine
transition triggered by 802.3ah CTRL message. For P2P, this will cause
the link to be always in DOWN state after the fiber is flapped. This CL
added the link update detection based on the receiver's optical signal
detection. This is only tentative solution until the wavelength
selection feature is ready.

Change-Id: I65b49190bc4a6d2a21145d732b562cfc36da42be