i2c: mv64xxx: separated transaction segments by repeated STARTs.
All transaction segments (all i2c_msg's passed to a single call to master_xfer)
should be separated on the bus by repeated START conditions. This change makes
that so with the exception of zero-byte reads (which don't make sense since the
slave will assume control of the data line). Zero-byte reads early terminate the
transaction.

Change-Id: I5ed9cbd94b6df0dc0bd33a7b8c0078e396fecd24