Printk and command line changes.
Increase the kernel buffer size for printk
to 1MB.

DISABLE! PRINTK_PERSIST, the kernel hangs at boot
when this is actually enabled so leaving it off.
Will look at what's going on with that later.

Copy how MIPS does command line handling.  If
CMDLINE_BOOL is set to y, then append the baked in
command line to the command line from the bootloader.
If CMDLINE_OVERRIDE is also set, then only use the
baked in command line.

Change-Id: Ia647a2e84ce44b6d684dd038626bba9b72ffcc7f