Marvell: MHSLPONLGCY-161 Wrong packet size when forwarding from mv_cph from WAN to LAN and vice-versa
"""
There was an old bug in the CPU packet handling driver (mv_cph) when
forwarding a broadcast packet from LAN to WAN and vice-versa which was
not removing the CRC from the original packet. The effect was the CRC of
the original packet was added to the payload and consequently the size
of the final packet was 4 extra bytes.
"""

Patch applied by replacing their directories with ours:
"Source/Kernel/linux_feroceon/arch/" -> "arch/"

In reference to b/33751084
Change-Id: If05ca347876676ca549e19d1b160537626375876