GFLT400: Syncing boardID values with uboot
When Marvell upgraded their SDK from RC32 -> thunderbolt_v1, they added a
new ID. We did not upgrade the uboot half. This means the IDs are
out of sync. Since uboot doesn't have this ID at all, we manually move
the extra one in the kernel to the end of list so that every other boardID
still lines up.

Change-Id: Ic557a092fb7e1e48cba91b11b67402d07679aa25